refactor: migrate agents off app.store.object_store
