+++ app/agents/chunker/agent.py
@@ -4,7 +4,7 @@ from datetime import datetime, timezone
 from typing import Any
 
 from app.events.types import INGEST_CHUNK_DONE
-from app.store.object_store import ObjectStore
+from app.objects import ObjectStore
 from app.services.audit import audit_event
 
 AGENT = "chunker"
