+++ tests/agents/panel_agent/test_panel_receipts.py
@@ -19,7 +19,7 @@ from app.agents.panel.writeback import AI_STATUS_HEADER, MAX_RECEIPTS
 from app.agents.panel_agent.agent import run_panel_intent_for_note
 from app.agents.panel_agent.runtime import PanelRuntimeResult, execute_panel_intent
 from app.store import object_store as object_store_module
-from app.store.object_store import DomainObject, ObjectStore
+from app.objects import DomainObject, ObjectStore
 
 
 class _StubReasoningFacade:
