+++ tests/agents/test_projector.py
@@ -8,7 +8,7 @@ import uuid
 from datetime import datetime, timezone
 
 from app.agents.projector.agent import run as project_run
-from app.store.object_store import ObjectStore, DomainObject
+from app.objects import ObjectStore, DomainObject
 
 os.environ.setdefault("DATABASE_URL", "postgresql+psycopg://app:app@127.0.0.1:15432/app")
